% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\title{Use the \href{https://github.com/CrossRef/rest-api-doc}{Crossref REST API}}
\description{
Use the \href{https://github.com/CrossRef/rest-api-doc}{Crossref REST API}
}
\section{Warning}{

If you are using this in your own package, or create a lot of traffic,
please set your own \code{\link[httr:user_agent]{httr::user_agent()}}.
}

\concept{api}
