% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{can_api}
\alias{can_api}
\title{Test whether API can be reached}
\usage{
can_api()
}
\description{
Test whether API can be reached
}
\seealso{
Other api functions: 
\code{\link{req_cr}()}
}
\concept{api functions}
