% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crlite-package.R
\docType{package}
\name{crlite-package}
\alias{crlite}
\alias{crlite-package}
\title{crlite: Lightweight R wrapper for the Crossref REST API}
\description{
Wraps the Crossref REST API for production uses.
    Always returns type and length-stable results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://subugoe.github.io/crlite}
  \item \url{https://github.com/subugoe/crlite}
  \item Report bugs at \url{https://github.com/subugoe/crlite/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Held \email{info@maxheld.de} (\href{https://orcid.org/0000-0002-4703-5388}{ORCID})

Other contributors:
\itemize{
  \item Göttingen State and University Library [copyright holder, funder]
}

}
\keyword{internal}
