% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pools.R
\name{polite}
\alias{polite}
\alias{get_cr_mailto}
\title{Accessing the polite API pool}
\usage{
get_cr_mailto()
}
\description{
Accessing the polite API pool
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{get_cr_mailto}: Get the email address to authenticate into the polite pool

In this order, returns the first hit of:
\enumerate{
\item git user email address for the repo at the working directory
(requires git to be configured),
}

Or errors out.
}}}

\seealso{
Other api pool access functions: 
\code{\link{pools}}
}
\concept{api pool access functions}
