% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pools.R
\name{pools}
\alias{pools}
\alias{resp_cr_pool}
\alias{was_pool}
\alias{can_pool}
\title{Diagnosing API pools}
\usage{
resp_cr_pool(resp)

was_pool(resp, pool = c("public", "polite", "plus"))

can_pool(pool = c("public", "polite", "plus"))
}
\arguments{
\item{resp}{An HTTP response object, as created by \code{\link[httr2:req_perform]{req_perform()}}.}

\item{pool}{giving the name of the expected pool.
Listed in order of increasing performance.}
}
\description{
Diagnosing API pools
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{resp_cr_pool}: Show the API pool which served a request
}
\itemize{
\item \code{was_pool}: Was the request served by the expected pool?
}
\itemize{
\item \code{can_pool}: Can the request be served by an expected pool?
}}}

\seealso{
Other api pool access functions: 
\code{\link{plus}},
\code{\link{polite}}

Other api pool access functions: 
\code{\link{plus}},
\code{\link{polite}}
}
\concept{api pool access functions}
