% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{req_ua_cr}
\alias{req_ua_cr}
\title{Set User Agent for HTTP requests}
\usage{
req_ua_cr(req, .mailto = getOption("crlite.mailto", default = NULL))
}
\arguments{
\item{req}{A \link[httr2]{request}.}

\item{.mailto}{a character string giving an email address to reach the user.
Passed on to Crossref as part of the user agent.
Used to contact the user if there are problems.
Defaults to \code{NULL}.
If provided, requests are served by a more performant
\href{https://api.crossref.org/swagger-ui/index.html}{polite pool}.}
}
\description{
Set User Agent for HTTP requests
}
\seealso{
Other api: 
\code{\link{req_cr}()}
}
\concept{api}
\keyword{internal}
