% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{req_user_agent_cr}
\alias{req_user_agent_cr}
\title{Set a user agent}
\usage{
req_user_agent_cr(req, mailto = NULL)
}
\arguments{
\item{req}{A \link[httr2]{request}.}

\item{mailto}{a character scalar giving a valid email address.}
}
\description{
Set a user agent
}
\keyword{internal}
