% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{verb_cr}
\alias{verb_cr}
\title{Use the \href{https://github.com/CrossRef/rest-api-doc}{Crossref REST API}}
\usage{
verb_cr(..., .contact = getOption("crlite.contact", default = NULL))
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[httr:VERB]{httr::VERB}}
  \describe{
    \item{\code{verb}}{Name of verb to use.}
    \item{\code{url}}{the url of the page to retrieve}
    \item{\code{config}}{Additional configuration settings such as http
authentication (\code{\link[httr:authenticate]{authenticate()}}), additional headers
(\code{\link[httr:add_headers]{add_headers()}}), cookies (\code{\link[httr:set_cookies]{set_cookies()}}) etc.
See \code{\link[httr:config]{config()}} for full details and list of helpers.}
    \item{\code{body}}{One of the following:
\itemize{
\item \code{FALSE}: No body. This is typically not used with \code{POST},
\code{PUT}, or \code{PATCH}, but can be useful if you need to send a
bodyless request (like \code{GET}) with \code{VERB()}.
\item \code{NULL}: An empty body
\item \code{""}: A length 0 body
\item \code{upload_file("path/")}: The contents of a file.  The mime
type will be guessed from the extension, or can be supplied explicitly
as the second argument to \code{upload_file()}
\item A character or raw vector: sent as is in body. Use
\code{\link[httr:content_type]{content_type()}} to tell the server what sort of data
you are sending.
\item A named list: See details for encode.
}}
    \item{\code{encode}}{If the body is a named list, how should it be encoded? Can be
one of form (application/x-www-form-urlencoded), multipart,
(multipart/form-data), or json (application/json).

For "multipart", list elements can be strings or objects created by
\code{\link[httr:upload_file]{upload_file()}}. For "form", elements are coerced to strings
and escaped, use \code{I()} to prevent double-escaping. For "json",
parameters are automatically "unboxed" (i.e. length 1 vectors are
converted to scalars). To preserve a length 1 vector as a vector,
wrap in \code{I()}. For "raw", either a character or raw vector. You'll
need to make sure to set the \code{\link[httr:content_type]{content_type()}} yourself.}
    \item{\code{handle}}{The handle to use with this request. If not
supplied, will be retrieved and reused from the \code{\link[httr:handle_pool]{handle_pool()}}
based on the scheme, hostname and port of the url. By default \pkg{httr}
requests to the same scheme/host/port combo. This substantially reduces
connection time, and ensures that cookies are maintained over multiple
requests to the same host. See \code{\link[httr:handle_pool]{handle_pool()}} for more
details.}
  }}

\item{.contact}{a character string giving an email address to reach the user.
Passed on to Crossref.
Used  to contact the user if there are problems.
Defaults to \code{NULL}.
If provided, requests are served by a more performant
\href{https://github.com/CrossRef/rest-api-doc#good-manners--more-reliable-service}{polite pool}.}
}
\description{
Use the \href{https://github.com/CrossRef/rest-api-doc}{Crossref REST API}
}
\concept{api}
