% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{req_cr}
\alias{req_cr}
\title{Use the \href{https://api.crossref.org/swagger-ui/index.html}{Crossref REST API}}
\usage{
req_cr()
}
\description{
Use the \href{https://api.crossref.org/swagger-ui/index.html}{Crossref REST API}
}
\seealso{
Other api functions: 
\code{\link{can_api}()}
}
\concept{api functions}
